// src/Models/LanguageId.ts
enum LanguageId {
  C = "c",
  CPP = "cpp",
  CSharp = "csharp",
  CSS = "css",
  FSharp = "fsharp",
  Go = "go",
  HTML = "html",
  Ini = "ini",
  Java = "java",
  JavaScript = "javascript",
  JavaScriptReact = "javascriptreact",
  JSON = "json",
  JSONWithComments = "jsonc",
  LaTeX = "latex",
  Less = "less",
  Lua = "lua",
  Makefile = "makefile",
  Markdown = "markdown",
  ObjectiveC = "objective-c",
  ObjectiveCPP = "objective-cpp",
  Perl = "perl",
  Perl6 = "perl6",
  PHP = "php",
  PowerShell = "powershell",
  Pug = "jade",
  Python = "python",
  R = "r",
  Razor = "razor",
  Ruby = "ruby",
  Rust = "rust",
  SCSS = "scss",
  SASS = "sass",
  ShaderLab = "shaderlab",
  ShellScript = "shellscript",
  SQL = "sql",
  Swift = "swift",
  TypeScript = "typescript",
  TypeScriptReact = "typescriptreact",
  TeX = "tex",
  VisualBasic = "vb",
  XML = "xml",
  XSL = "xsl",
  YAML = "yaml",
}

export default LanguageId;